fix: modularize ModalStateManager
