+++ frontend/src/components/SeedlotRegistrationSteps/OrchardStep/styles.scss
@@ -28,12 +28,12 @@
 }
 
 #female-gametic-combobox::placeholder {
-  color: var(--bcgov-text-primary);
+  color: var(--#{vars.$bcgov-prefix}-text-primary);
   opacity: 1;
 }
 
 .male-gametic-invalid .#{vars.$bcgov-prefix}--radio-button + .#{vars.$bcgov-prefix}--radio-button__label .#{vars.$bcgov-prefix}--radio-button__appearance {
-  outline: 2px solid var(--bcgov-support-error);
+  outline: 2px solid var(--#{vars.$bcgov-prefix}-support-error);
   outline-offset: 1.5px;
 }
